Just a small Chnage to the Comments and The Way They Are Displayed
i added a "file Tag" to the comments, because we are having some problems with that aspect of the project, so maybe that's a way around it. in Short, the comments appear in the following format
File Name: (fileName). Line: (line), (commentType).
P.S Dont worry, this is very easy to revert